added put route to comment_router

+++ app/api/comment_routes.py
@@ -31,13 +31,26 @@ def post_comment():
     return comment.to_dict()
   return {'errors': validation_errors_to_error_messages(form.errors)}, 401
 
-@comment_routes.route('/<int:id>', methods=['DELETE'])
+@comment_routes.route('/<int:id>', methods=['DELETE', 'PUT'])
 @login_required
 def delete_comment(id):
   comment = db.session.query(Comment).get(id)
-  if current_user.id == comment.user_id:
-    db.session.delete(comment)
-    db.session.commit()
-    return {'message': 'Comment deleted'}
-  else:
-    return {'message': 'Unauthorized'}
+  if request.method == 'DELETE':
+    if current_user.id == comment.user_id:
+      db.session.delete(comment)
+      db.session.commit()
+      return {'message': 'Comment deleted'}
+    else:
+      return {'message': 'Unauthorized'}
+  elif request.method == 'PUT':
+    form = CommentForm()
+    form['csrf_token'].data = request.cookies['csrf_token']
+    if form.validate_on_submit():
+      comment = Comment(
+        project_id=form.data['project_id'],
+        user_id=form.data['user_id'],
+        description=form.data['description'],
+      )
+      db.session.commit()
+      return comment.to_dict()
+    return {'errors': validation_errors_to_error_messages(form.errors)}, 401
